refactor: move stage type
